versao3/pyproject.toml
[project]
name = "newsroom-v3"
version = "0.1.0"
description = "Brasileira.News V3 modular multi-agent newsroom"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Brasileira News", email = "dev@brasileira.news" }]
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "redis>=5.0",
  "psycopg[binary]>=3.1",
  "feedparser>=6.0",
  "beautifulsoup4>=4.12",
  "aiokafka>=0.10.0",
  "lxml>=5.1.0",
  "tenacity>=8.2.3",
  "numpy>=1.26.4",
  "kafka-python-ng>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
